Add the ability to pass a un/pw to fetch_rss
This change adds an explicit user and pass parameter to fetch_rss, to be used
when fetching the rss feed from the feed url.

Passing user/pass on the url works as an alternative to this, but only if your
credentials don't contain characters that cause php's parse_url to fail - a '/'
is one such character.

Closes: https://github.com/kellan/magpierss/pull/11